refactor data models and update feature engine
